Move to allowlist configuration for rules
Summary:
One of the big hassles with upgrading RuboCop is figuring out which new lints
we're going to keep, while also looking for regressions on existing cops. By
switching to an allowlist approach, we can update to new versions and check for
regressions on existing lints first, and then ratchet in the new lints as time
permits.

Reviewed By: lmaghran

Differential Revision: D38557880

fbshipit-source-id: dcfaedbeac36f7aaccc0176413267dec466fecca