Support multiple path definitions for single external plugin package
Summary:

External Plugins need to be define as `[name="package", path="location"]` however this approach attach to a relation one package one library. For distribution purposes this approach does not work for when a single package distributes multiple plugins/libraries.

To extend External Plugins support to allow multiple paths for a single package we:
 1. Made the `path` key to receive a string or a list.
 1. Backwards compatibility is ensure validating the format as a string value, if so then the value of `path` is passed
 1. If the `path` value is not an string we evaluate as if it is a list and iterate over it.

Reviewed By: dafyddcrosby

Differential Revision: D38633473

fbshipit-source-id: a86af08dcf6205ba3c59ce0e46999350cb604822