fb_timers: add FixedRandomDelay=true to make splays more predictable
Summary:
`FixedRandomDelay=` was added to timers in systemd 247 by https://github.com/systemd/systemd/pull/17497. The goal is to make `RandomizedDelaySec=` more consistent by using a constant delay value based on the timer and features of the host. Otherwise, the random delay is recalculated whenever `systemctl daemon-reload` is run or the timer is restarted. `FixedRandomDelay=` has no effect unless `RandomizedDelaySec=` is set to non-zero.

This fixes a problem we've had where timer iterations were skipped because the delay was being recalculated further into the future every time `systemctl daemon-reload` was called.

Reviewed By: joshuamiller01

Differential Revision: D38056851

fbshipit-source-id: 48ffb387177d4823ad9543bb68959028e0b0165c