Make inability to determine the root device explicitly incompatible with the chef storage API in fb_storage
Summary:

This change is an enhancement to the fb_storage::devices_to_skip to raise an exception when the root device is not found. It's currently a safe change as default value is strict = false, hence old functionality is still applied and just prints a log message if the root device is not identified.

Reviewed By: joshuamiller01

Differential Revision: D40216286

fbshipit-source-id: 28099074dc619fdbda567f64d08d11ab696c3fa9