Remove unnecessary start of FluentBit service
Summary:

FluentBit on Windows `start`  and then `restart` at the end of chef run. This is duplicating the start action in Windows making a race condition where the service `start`s, then the delayed restart triggers a `stop` and then `start`. The problem comes in the `stop`. The stop action is triggered right after the initial `start` provoking that in windows the FluentBit service may not have finished its start process while receiving a stop signal. Due to internal windows resource management we have observed that if the service is not fully started and try to stop it the process will hung in a Stopping state similar to zombie in a never ending stopping.

To fix previous when happens in Chef it requires manual intervention to kill the Windows task. To prevent this race condition we are deleting the unecesseray start. As Fluentbit will always trigger restart on every install/upgrade as the cookbook will modify the config files triggering a delay

**Show how on windows it triggers double start**
{F785199547}

**Shows the error happening**
{F785199686}

Reviewed By: joshuamiller01

Differential Revision: D40610771

fbshipit-source-id: 28578bb96d2bcf7be9e018b61c10325eaebebab0