Enable attestation and TPM on aarch64
Summary:

We're far enough along that we can ungate these service. Note that attestation is still gated in the backend, so this is a functional no-op for now.

NOTE: We ensure that dbus-tools is installed, as this is needed by this recipe and not otherwise brought in on the CentOS 9 (at least on aarch64) hosts.

Reviewed By: joshuamiller01

Differential Revision: D41337124

fbshipit-source-id: 99c8daf9f310ce0f712ab9843ce44c2629a6c37d