Allow usage of custom restart scripts for the fluentbit service
Summary:

FluentBit service can be restarted by default stop/start actions of Service Registry. However there is no option to run a custom process or script to restart the service.

On this commit we are adding the functionality to define a custom restart script for restart of the windows service.

## How to add a custom restart script
1. The cookbook that consumes `fb_fluentbit` should deploy its custom script first
1. Define the attribute `node.default['custom_svc_restart_command']` to the script filepath previously created

Reviewed By: gogsbread, leoswaldo

Differential Revision: D41966104

fbshipit-source-id: 6d1984b9b5c99f682df050f5a9ee9fbc1ac6547e