Update windows fb_tmpclean to correctly use _hours_ instead of _seconds_ by default
Summary:

This was a function of me misreading the readme - for whatever reason, I thought the time value
was days, but it's hours.  This also addresses a case which only affects some earlier windows,
where the glob of the folder was returning the contents in some cases, but the directory in the other
case.  This now works for both on all windows versions, not just 2019+ where the behaviour was changed (before it was ineffective on 2016 - not dangerous, would just not do anything unelss the pathname was a glob.  so c:\windows\tem* worked, but c:\windows\temp would do nothing on that platform)

This also includes the adding of -File to the powershell line in the task, as it generally should have...
it actually works without it, but it's more correct, so we'll use this oppertunity to add it

Reviewed By: joshuamiller01

Differential Revision: D42037626

fbshipit-source-id: fcade647d5f6fb752056ee8c0acdad63da1e5668