don't look at inode64 opt for tmpfs when comparing opts with kernel >= 5.9
Summary:

tmpfs opts always have differences detected by fb_fstab recipe. This seems to be linked to the fact that since kernel 5.9, tmpfs are mounted with inode64 opt - even if not metionned in `/etc/fstab`.
This result in RepeatingResources, but thoses are false positive.
The fix consists in removing `inode64` opt when comparing `desired` and `mouted` opts WHEN kernel is >= 5.9

Reviewed By: Zepheus

Differential Revision: D39275841

fbshipit-source-id: a4bfefe1ce14c4a817bfcee4d5c6d71f6e387999