stop installing the virtual 'iptables' package on centos9
Summary:

On CentOS Stream 9 (and all supported Fedora releases) `iptables` is now a virtual package, provided by `iptables-nft` (in c9s) and `iptables-legacy` (in Fedora, and in EPEL 9 via the `iptables-epel` package, since CentOS Stream dropped support for the legacy backend).

Since this is the one we want to use, on CentOS 9 we should install `iptables-legacy` instead of the virtual package. This will also restore idempotency, as when asking Chef to install a virtual package, it will *always* execute the `package` resource because the package never actually gets installed (but one of the implementing package does. We got lucky that `legacy` sorts before `nft` so we accidentally get that installed)

Also clean up the Fedora code which was:
- specifying specific Fedora releases (and was not updated for 37)
- using -compat which is the transition package (which also pulls in iptables-utils which is not needed)

And drop the code path handling CentOS 6 which is long EOLed

Reviewed By: joshuamiller01

Differential Revision: D42905172

fbshipit-source-id: ab2ca739e025602ba1e0ae4d25127fed8f8205c3