fix(fb_apache): render arrays inside of hash configs
This commit moves all template logic into a single function and drops
the separate `apache_conf.erb`, which should prevent things from only
working in one context (e.g. at the top-level of the config) and not
another (e.g. as a sub-hash of the config).

I replaced the check for `Fixnum` with `Integer` when moving the logic,
per the `Lint/UnifiedInteger` lint.