Add minimum and maximum supported OS release methods to node
Summary: Right now we have a lot of places where we have gating on something like `node.centos8?`. Many of these lines have been added under the assumption that C8 is the latest OS. The net effect of this has been that sometimes CentOS9 is falliing back to legacy C7 behaviour instead of using the behavior that would work correctly for C8.

Differential Revision: D44666642

fbshipit-source-id: 60d35ccc706732f2ff0f0026f1c8a7c586637230