fb_timers: refactor autostart handling for old systemd versions
Summary:

Refactor autostart handling for old systemd versions.  This makes the code
easier to read in differentiating when we're setting up timers the normal way
(with modern systemd) versus the old way, and will make future transitioning to
a different model (forthcoming) simpler also.

This should be a no-op; node['packages'] is ohai-time and reordering the evaluation
within a custom resource should be irrelevant.

Reviewed By: dafyddcrosby

Differential Revision: D45199845

fbshipit-source-id: 12b6ab20c90c59d1c9660f411de9fb29f2b295c4