fb_timers: CPU perf improvements via avoiding read-write loops and node.rm
Summary:

If during a loop you modify an element of the node object, and then read from the same ImmutibleHash,
it causes some kind of overhead which costs non-trivial amounts of CPU when iterating enough times.

In other words: if you modify node.default['something'][...], then read from node['something'], and then
loop and do that again and again, you get significant perf overhead.

If the read is removed, this overhead is avoided; this diff takes an unnecessary read of `node['fb_timers']['_timer_path']`
and captures it in an lvar, which saves ~5ms per iteration in my environment.

While we're here, also switch from node.rm to node.rm_default, because that's also much less CPU expensive.

Reviewed By: dafyddcrosby

Differential Revision: D45459381

fbshipit-source-id: c641c67a3a6868a6bccb63d4b507a1d510295dad