Add whole disks instead of partitions in `devices_to_skip`
Summary:

In `storage.rb` there is `devices_to_skip` which is filled with all the devices to skip, which are the ones for / and /boot.
If / is mounted from a RAID array, all the devices in the array (e.g. /dev/sda1 and /dev/sdb1) are added to devices_to_skip. But when doing so, partitions are added, not disks.

While at at:
- fix device_name_from_partition() which was unable to handle loop devices and few others
- fix typos in comments
- add more tests

Differential Revision: D45199010

fbshipit-source-id: 7f94df0b70ae416c639928255f6b4b401068adf0