Allow mount point exclusions by regex
Summary:

The `.snapshot` directory on the NetApp is a special directory that only gets mounted when its accessed. Since we have `enable_unmount=true`, Chef will attempt to unmount the volume and fail if its in use. Currently we can only ignore by prefix, I decided to add an ignore by regex to give a bit more control vs doing something an include.

Differential Revision: D46297189

fbshipit-source-id: 4e5c614218c271d5c8df9883cb754faf0e5abd91