Delete fb_smartctl cookbook
Summary:

**Internal:**
With the cleanup of removing `/opt/homebrew`, this helps removes the symlink to the path since it no longer exists.

This gets rid of the need for the symlink of `/opt/homebrew/bin/smartctl` plus gets rid of the 2014 Mac Mini hardware.

There shouldn't have any (CPE-specific) use case for 2014 mac minis, so any code explicitly targeting them is almost certainly a safe candidate for summary removal.

As someone once stated, "*they should be lovingly and affectionately hit with a blackjack and have it confiscated. :)*"

Plus, `fb_macos_shared::smartmontools` recipe already handles the nix version of this tool, making this cookbook redundant.

Differential Revision: D46707384

fbshipit-source-id: c6475554f3c878270fa760a9b4ad8df377d532d3