Adopt fluent-bit name by default instead of td-agent-bit
Summary:

FluentBit changed the name of its binaries from `td-agent-bit` to `fluent-bit` starting 1.9 versions. Therefore we needed to have logic that supported both names for a period of a time. It has been over 6 months since this operation and we are moving to default to `fluent-bit` for name.

The logic to support `td-agent-bit` name will continue to be supported for the coming 6 months, until April 15th, 2024.

Differential Revision: D50092325

fbshipit-source-id: 3d9c5a944728971a18a15923e831b1b1a2ca8fe4