fb_systemd: Don't call systemd-sysusers on older versions of systemd
Summary:

D36004124 updated the `fb_systemd` recipe to call `systemd-sysusers`
to create a missing user. However, the command uses the `--inline` flag that
was only added in systemd 238. D36528947 was sent to update the minimum version
to 215 (which is when the `system-sysusers` command was added). However, the
`--inline` flag was still missing from that version. Thanks to gmambro for
spotting that!

From the systemd changelog: https://github.com/systemd/systemd/blob/main/NEWS

```
CHANGES WITH 238:

<snip>

 * systemd-sysusers gained a mode where the configuration to execute is
          specified on the command line, but this configuration is not executed
          directly, but instead it is merged with the configuration on disk,
          and the result is executed. This is useful for package installation
          scripts which want to create the user before installing any files on
          disk (in case some of those files are owned by that user), while
          still allowing local admin overrides.
```

The missing arg caused S372980 as the lego-linux tier runs Ubuntu VMs with
systemd 237. There's some more context in my post in the systemd group:
https://fb.workplace.com/groups/systemd.and.friends/posts/3509161785966261.

I think that it might be possible to remove this block of code entirely,
however I want to make the smallest change possible to fix lego-linux. Those
supporting systemd will have a better idea of whether the code can be removed.

Differential Revision: D50494316

fbshipit-source-id: db7bcff80672673aac48be372d2434b768c99d96