fb_networkd: fix boot service ordering to avoid races between udev, chef, and network-online.target
Summary:

There are several problems during boot right now wrt networkd, udev, chef, and the network-online.target
- networkd starts while the udev rules are running, and may try to apply rules that might be incorrect
- the udev rules are `networkctl reconfigure`ing interfaces at the same time networkd is trying to configure them
- chef-boot runs while the udev rules are running, and may observe interim interfaces or interfaces with the wrong MAC
- network-online.target doesn't actually block until the network is online; the target is reached before any interface is configured
  - this means that chef is also concurrently running while networkd is configuring the interfaces, basically creating the same problem as with the udev rules

The only reasonable way to tackle this is to make things execute in a proper order and end the races.
This does the following:
- make the udev rules not call `reconfigure`; these rules should only run on boot, so the interfaces should subsequently be configured by networkd
- make networkd wait for udev rules to 'settle', so that before networkd runs the interfaces have been renamed
- fix network-online.target to actually work - since chef already waits for network-online.target, if it actually works
there shouldn't be a problem

The end result is that the udev rules run, then when they're done, systemd-networkd starts up, and chef only runs once network-online.target is reached

Note: network-online.target is not "all interfaces are up and fully configured".  See https://www.freedesktop.org/software/systemd/man/latest/systemd-networkd-wait-online.service.html
This _also_ causes problems, but ... I'm not going to try to address that here.

Differential Revision: D50518948

fbshipit-source-id: 8daabb31555ae019b6f32867bcae4a2bc9ce8967