Fix chocolatey.exe detection
Summary:
This diff is a part of bigger effort to migrate from engenv chef to prod chef solo.

In the lab environment, there is no Chocolatey present on the machine during the chef bootstrap, and it's important to properly detect the `chocolatey.exe`.

In the recent chef versions (17.4.38) the behavior of the `which()` function changed, and now it's returning `false` instead of `nil`.
If you look at D36651569, you can see that the return value of `which()` is changing.
To be on the safe side, I am adding the test that looks for **both** `nil` and `false`

Differential Revision: D56729989

fbshipit-source-id: 35f071a9b2ab131dfa5ac71cff1bd7939d921864