special partition size handling for FBPN 01-005817, 01-005959
Summary:
eag-prod-2 has three different kinds of T16, each with slightly different storage capacities.

This diff makes such that each of the smaller capacity machines will format its disks to have
partitions passed to the VM that are the same size as the largest machine.

Confirmed we currently have only one of these machines for each FBPN. Other anvils
should not be impacted.

Had to add "s" to the allowed suffixes that can be passed for partition start/end in fb_storage
so that we can be precise in partition sizing.

Differential Revision: D58845625

fbshipit-source-id: a63b8fb28170053845e66423d27e15d89cd40bec