Facebook: [hypernode][oly2] nvme format with T10-DIX enabled if supported by device
Summary:
We want to enable T10-DIX support during nvme format for drives that support it. As of today, we were doing it only for a list of specific model Ids (currently all OLY2.0). As different models for OLY2.x are added to the fleet, we need to keep updating this list with model Ids that support T10-DIX.

This diff checks if a drive supports t10dix in the way we want and enables it if it does.

Differential Revision: D64211281

fbshipit-source-id: 17c6a0be38b9c8cc15cfef5e7b632334099e66b4