Update to the cookbook syncer
The syncer would silently overwrite your cli or env options
with a config file in a way that was confusing.

This properly handles an expected hierarchy of:
default -> config -> env -> cli

Also, the vim modeline was set to 2-space indentation, but the code
was 4-space, so update the code to match the vim modeline.

Signed-off-by: Phil Dibowitz <phil@ipom.com>