Update to the cookbook syncer (#244)
Summary:
The syncer would silently overwrite your cli or env options
with a config file in a way that was confusing.

This properly handles an expected hierarchy of:
default -> config -> env -> cli

Also, the vim modeline was set to 2-space indentation, but the code
was 4-space, so update the code to match the vim modeline.

Signed-off-by: Phil Dibowitz <phil@ipom.com>

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/244

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

Note: we don't use this script internally, it's for the benefit of our downstream open-source users (like Phil)

Differential Revision: D65949355

fbshipit-source-id: e49c3f519bdc0e80dfd3fa0c492139775ac86ed1