new cookbook: fb_smartmon (#203)
Summary:
This is a very simple cookbook to install and configure smartmontools.

This was primarily written by infcted

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/203

Test Plan: Potentially usable internally (we don't yet have a dedicated smartmon cookbook beyond engenv_smartd FWICT, but for now this would be an example on our open-source repo and not in use.

Differential Revision: D66516318

fbshipit-source-id: d1742fa753665cfdecd317354f0d69cd7f8d5005