Move library files to lib/
Summary:
This gets the cookbook tree in line with the bookworm-specific github
repo, while still working when run within the cookbook

Instead of using require_relative for libraries, we'll add the bookworm libraries to the LOAD_PATH in the wrapper script created in the prior diff in the stack. This allows us to follow the idiomatic lib/ directory layout seen in most Ruby gems (and used in the
Bookworm github repo).

Differential Revision: D68423773

fbshipit-source-id: 3c9acd5a1ffd971c9f0c55043486639c8b2e032c