fix(fb_apache): render arrays inside of hash configs (#218)
Summary:
## Description

This PR fixes a bug where the `template_hash_handler` in `fb_apache` silently skips over `Array` values. Now these values are rendered as if they were in the top-level `VirtualHost` directive.

For example:

```ruby
{
  ...
  'Location /foo' => {
    'Header' => [
      'unset Bar',
      'unset Baz',
    ],
    ...
```

...will now properly include multiple `Header ...` lines in the `Location /foo` directive.

## Impact

These values were ignored previously, so it is possible that it would cause changes in hosts where this was happening.

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/218

Differential Revision: D45869664

fbshipit-source-id: 947ec76843416d411d4288c23bba4125de163185