new cookbook: fb_networkmanager (#127)
Summary:
This cookbook aims to manage NetworkManager in the most controlled
way possible. It might-could-should be sent over to CPE cookbooks,
I can redirect over there if preferred.

While, much to my dismay, some people do run NM on servers, and this
will work for that, it was built to handle this client machine case.

NetworkManager treats its config files as databases and can change them,
so they can't be handled simply as templates - more finesse is needed.

However, write-once doesn't provide the level of management that
FB-API-style cookbooks expect. This marries the two. All desired
settings are managed 100%, and NetworkManager is allowed to make
additive settings (such as "seen-bssids" or "password" [sic]) so that it
can continue to function for users.

It is immune to the format changes that NM can make - spacing and
ordering changes will not cause the file to be updated any more than
ignorable settings. When updating the file, care is taken to stick
close to NM's style.

I considered making blacklist/whitelists of fields that NM could
modify, but in practice since we merge in the user's config, the user's
config is effectively a blacklist, and whitelist would likely be far
too fragile.

It has full docs, including on it's Migration Feature which will be
useful for anyone who had their own Chef to do this but wants to
migrate.

It also comes with pretty good test coverage.

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/127

Differential Revision: D69685647

fbshipit-source-id: 45632b1015565a87cc515f2b74d3dd15a6cf9894