New cookbook: fb_bind (#273)
Summary:
This cookbook manages Bind. It's got multiple ways of handling
zonefiles to accomodate shops that generate their own outside of Chef,
while providing a nice node-based API and helper functions for others.

Like every other cookbook it's very opinionated on where it puts it's
config, but not on the config itself. This helps with cleanup of
unowned cruft.

Signed-off-by: Phil Dibowitz <phil@ipom.com>

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/273

Differential Revision: D69810712

fbshipit-source-id: 9fd87179861eb068a4578ecfd2310065e04c3eae