Some fb_postfix fixups for Debian/Ubuntu
There are a few of our defaults that weren't correct on Debian/Ubuntu. I
suspect these are the root of what @PatrickDale was running into in #225.
I didn't see it because my local cookbooks overwrote these three
variables.

However, I'm rolling this out to more Debian systems and bumped into it,
so fixing it.

Signed-off-by: Phil Dibowitz <phil@ipom.com>