Fix `run_cookstyle`
Recent changes from #296 broke the ability to run `run_cookstyle` from
directories other than the root directory.

That, in-turn breaks the ability to run it in other repos, and I have
several repos that just call `run_cookstyle` from the checkout of this
repo to keep style in-sync.

This fixes both of those cases.

Signed-off-by: Phil Dibowitz <phil@ipom.com>