fb_networkmanager: fix missing default (#302)
Summary:
This got lost somewhere in the review process updates.

We name the files `fb_networkmanager_*` so that we can make sure we
only mess with things we own, but that's not what the user expects
the network to show up, so we must set `connection.id` to the name of
the network (unless the user has chosen something else).

Further, without doing this, NetworkManager will create a separate
connection for the network if this one has not yet been activated but
the network is seen.

Signed-off-by: Phil Dibowitz <phil@ipom.com>

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/302

Differential Revision: D72402056

fbshipit-source-id: 8b21263aad05ac9a2d084db2121100f5d5bd283d