Fix ChefUtils scoping
In an attempt to fix internal scoping conflicts, I scoped calls to
ChefUtils helpers, but you can't do that because they need to be mixed
in the the current DSL.

This drops the two scopes that were left. For `kitchen?` there's no
duplicate, so we're fine. For `debian?`, I moved to
`node.debian_family?`

Signed-off-by: Phil Dibowitz <phil@ipom.com>