fb_spamassassin: Fix timer jobs (#306)
Summary:
Somehow I had gotten it in my head when I wrote a bunch of these
cookbooks that timer jobs needed to `:enable`'d but `:start` would
run the actual job. That is, of course, not true, the underlying
_service_ can be started to do an immediate run.

As a result the update jobs never ran, when expected. This fixes that.

Signed-off-by: Phil Dibowitz <phil@ipom.com>

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/306

Differential Revision: D73198736

fbshipit-source-id: 6a50a4b2aa87eb8bc1382ed8f3728d4129b8cef6