Fix ChefUtils scoping (#305)
Summary:
In an attempt to fix internal scoping conflicts, I scoped calls to
ChefUtils helpers, but you can't do that because they need to be mixed
in the the current DSL.

This drops the two scopes that were left. For `kitchen?` there's no
duplicate, so we're fine. For `debian?`, I moved to
`node.debian_family?`

Signed-off-by: Phil Dibowitz <phil@ipom.com>

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/305

Differential Revision: D73197033

fbshipit-source-id: 304cd4aa90ff8ccda71b5495be2f0ed2811f28ad