add multi-parser support, AllJsonRecipes report
Summary:
With the addition of [JSON recipes in
Chef](https://github.com/chef/chef/pull/15094), we'll need to make Bookworm
aware of other file types besides Ruby.

For this diff, I'm focusing mostly on adding JSON parsing (ie I'll be updating
the built-in rules+reports to handle JSON recipes in a subsequent diff).

Differential Revision: D79741933

fbshipit-source-id: f8b3d4a1aa6667ca6ddd13267112a25f304e3e1f