fb_networkd: do not run networkctl reconfigure if an interface doesn't exist yet.
Summary:
The current issue with https://www.internalfb.com/diff/D78911435 is that:

> When the network and link files get generated, this gets called. The problem is there is no guarantee the interfaces exist since the driver may or may not be loaded. So when you call networkctl reconfigure bethX it fails. Same with udev trigger ....

I guess we could just have a guard that makes sure ohai actually saw the interface exist before we're trying to shell out and reconfigure.

Fair warning: I don't have a lot of experience with how we configure out interfaces in chef, so this is more of an attempt to get a discussion going.

Differential Revision: D79558360

fbshipit-source-id: ec385b85fca4ad345acd1038ebebae401da7a185