allow opt-out for override directory management
Summary:
Current code cleans up the override directory for delete action if
it's empty. This only makes sense if the override directory is exclusively
managed by fb_systemd_override, which is not always the case. Add a flag to
allow leaving the override directory alone so later setup won't have to
recreate it.

Differential Revision: D85810625

fbshipit-source-id: 6fbca1772805a767421279058a3aefde12867389